requirements.txt
Django==3.0
mysqlclient
Pillow
argparse
diff-match-patch
django-date-extensions
django-import-export
django-mysql
django-queryset-csv
django-tinymce
et-xmlfile
jdcal
odfpy
openpyxl
simplejson
six
tablib
#wsgiref
xlrd
